updated tables

+++ tables/transaction.sql
@@ -1,16 +1,10 @@
--- Author: Tyler Pena
--- Created: 2/25/21
--- Creates Transaction Table
-
-use opdex;
-
 create table transaction
 (
     Id          bigint          not null,
     `From`      varchar(50)     not null,
     `To`        varchar(50)     not null,
     Hash        varchar(50)     not null,
-    GasUsed     smallint        not null,
+    GasUsed     int             not null,
     Block       bigint unsigned not null,
     CreatedDate datetime        not null,
     constraint transaction_Id_uindex
